pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "orqestra"
version = "0.1.0"
description = "Business strategy agent with integrated knowledge base"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "python-multipart>=0.0.9",
    "python-frontmatter>=1.0",
    "trafilatura>=1.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "python-telegram-bot>=21.0",
    "telegramify-markdown>=1.0,<2",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sse-starlette>=2.0",
]

[project.optional-dependencies]
analysis = [
    "pandas>=2.0",
    "matplotlib>=3.7",
    "openpyxl>=3.1",
    "numpy>=1.24",
]
browser = [
    "playwright>=1.40",
]
test = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
orqestra = "orqestra.main:main"
orqestra-api = "orqestra.gateway_api:main"
orqestra-telegram = "orqestra.gateway_telegram:main"
